requirement.txt
arrow==0.7.0
defusedxml==0.4.1
Django==1.9
django-social-auth==0.7.28
gunicorn==19.4.5
httplib2==0.9.2
oauth2==1.9.0.post1
oauthlib==1.0.3
PyJWT==1.4.0
pyserial==3.0
python-dateutil==2.4.2
python-openid==2.2.5
python-social-auth==0.2.13
python3-openid==3.0.9
pytz==2015.7
requests==2.8.1
requests-oauthlib==0.5.0
six==1.10.0
wheel==0.24.0
